Move URI template from constants into service
The URI templates are tightly coupled to the parameters the template
accepts. The theory for centralizing all the templates into the
constants.js file was that it would make refactoring easier. It's very
unlikely that we will ever refactor the URL space where we would update
the space for more than one service.

Issue: #157